+++ src/main/java/org/oceanops/api/orm/auto/_GtsRecorderType.java
@@ -3,7 +3,6 @@ package org.oceanops.api.orm.auto;
 import java.io.IOException;
 import java.io.ObjectInputStream;
 import java.io.ObjectOutputStream;
-import java.math.BigDecimal;
 import java.util.List;
 
 import org.apache.cayenne.BaseDataObject;
@@ -27,13 +26,13 @@ public abstract class _GtsRecorderType extends BaseDataObject {
 
     public static final StringProperty<String> CODE = PropertyFactory.createString("code", String.class);
     public static final StringProperty<String> DETAILS = PropertyFactory.createString("details", String.class);
-    public static final NumericProperty<BigDecimal> ID = PropertyFactory.createNumeric("id", BigDecimal.class);
+    public static final NumericProperty<Integer> ID = PropertyFactory.createNumeric("id", Integer.class);
     public static final StringProperty<String> TYPE = PropertyFactory.createString("type", String.class);
     public static final ListProperty<Obs> OBSS = PropertyFactory.createList("obss", Obs.class);
 
     protected String code;
     protected String details;
-    protected BigDecimal id;
+    protected Integer id;
     protected String type;
 
     protected Object obss;
@@ -58,12 +57,12 @@ public abstract class _GtsRecorderType extends BaseDataObject {
         return this.details;
     }
 
-    public void setId(BigDecimal id) {
+    public void setId(Integer id) {
         beforePropertyWrite("id", this.id, id);
         this.id = id;
     }
 
-    public BigDecimal getId() {
+    public Integer getId() {
         beforePropertyRead("id");
         return this.id;
     }
@@ -127,7 +126,7 @@ public abstract class _GtsRecorderType extends BaseDataObject {
                 this.details = (String)val;
                 break;
             case "id":
-                this.id = (BigDecimal)val;
+                this.id = (Integer)val;
                 break;
             case "type":
                 this.type = (String)val;
@@ -163,7 +162,7 @@ public abstract class _GtsRecorderType extends BaseDataObject {
         super.readState(in);
         this.code = (String)in.readObject();
         this.details = (String)in.readObject();
-        this.id = (BigDecimal)in.readObject();
+        this.id = (Integer)in.readObject();
         this.type = (String)in.readObject();
         this.obss = in.readObject();
     }
